+++ macros/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Adapt to the new `Serialize` and `Deserialize` traits.
 - Rename the generated enum for function calls from `[..]Function` to `[..]Call`.
+- In `generate! {}`, the `crate` argument now takes a path like `my_reexports::my_aldrin` instead of
+  a literal string.
 
 ## [0.12.0] - 2025-01-26